Implement TcpEstablished.Transmit
I had to unseal the `TcpConnection` class in order to enable a Test Spy.
This is in itself okay, I think, since I don't think that Design
Pattern's example class is sealed. As a matter of fact, I can't even
remember if that's something one can do in C++...

In any way, it's a slight deviation from the book's example code to make
the `ProcessOctet` method virtual, but I don't think that this changes
the example's power to demonstrate the pattern structure.